backend/requirements.txt
Django>=3.0,<4.0
django-unixdatetimefield
psycopg2-binary>=2.8
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
